requirements.txt
# Core ML libraries
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.6.1
scipy==1.12.0

# Visualization
matplotlib==3.10.0
seaborn==0.13.2
plotly==5.19.0

# Deep Learning
# tensorflow==2.18.0 # Temporarily commented out
keras==3.8.0

# Jupyter environment
jupyterlab==4.3.5
ipython==8.21.0
ipykernel==6.29.5
ipywidgets==8.1.5

# ML utilities
openml==0.15.1
mlxtend==0.23.4
imbalanced-learn==0.13.0
category_encoders==2.8.0
joblib==1.3.2

# Data handling
gdown==5.2.0
GPy==1.13.2
graphviz==0.20.3

# Web framework
streamlit==1.32.0
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.3
requests==2.31.0
python-dotenv==1.0.1
mysql-connector-python==8.3.0
